streamlit_client/pyproject.toml
[project]
name = "jira-cs-streamlit-client"
version = "1.0.0"
description = "Streamlit 客戶端 for Jira CS Server - 基於 LangGraph + LangMem 的智能客服系統"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.30.0",
    "requests>=2.31.0",
    "psycopg2-binary>=2.9.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
